feat: create new empty alembic file that points to prod revision hash
