--- backend/alembic/versions/2025_12_17_0915-861356fccd4d_init.py
@@ -1,102 +1,0 @@
-"""init
-
-Revision ID: 861356fccd4d
-Revises:
-Create Date: 2025-12-17 09:15:46.363604
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision: str = "861356fccd4d"
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "user",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("clerk_id", sa.String(), nullable=False),
-        sa.Column(
-            "contact_number",
-            sqlalchemy_utils.types.encrypted.encrypted_type.EncryptedType(),
-            nullable=True,
-            comment="Assumes SG phone number",
-        ),
-        sa.Column("created_at", sa.TIMESTAMP(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("clerk_id"),
-    )
-    op.create_table(
-        "care_receipient",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column(
-            "name",
-            sqlalchemy_utils.types.encrypted.encrypted_type.EncryptedType(),
-            nullable=False,
-        ),
-        sa.Column(
-            "contact_number",
-            sqlalchemy_utils.types.encrypted.encrypted_type.EncryptedType(),
-            nullable=False,
-            comment="Assumes SG phone number",
-        ),
-        sa.Column(
-            "alias",
-            sa.String(),
-            nullable=False,
-            comment="To prevent data overflow on frontend",
-        ),
-        sa.Column("app_language", sa.String(), nullable=False),
-        sa.Column("age", sa.Integer(), nullable=False),
-        sa.Column("race", sa.String(), nullable=False),
-        sa.Column("gender", sa.String(), nullable=False),
-        sa.Column("postal_code", sa.Integer(), nullable=False),
-        sa.Column("floor", sa.Integer(), nullable=False),
-        sa.Column("block", sa.String(), nullable=False),
-        sa.Column("unit", sa.String(), nullable=False),
-        sa.Column("consecutive_checkins", sa.Integer(), nullable=False),
-        sa.Column("is_suspended", sa.Boolean(), nullable=False),
-        sa.Column("created_at", sa.TIMESTAMP(), nullable=False),
-        sa.Column("user_id", sa.Integer(), nullable=True),
-        sa.Column("can_record_mood", sa.Boolean(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["user.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "mood",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("care_receipient_id", sa.Integer(), nullable=True),
-        sa.Column(
-            "mood", sa.Enum("HAPPY", "OK", "SAD", name="selectedmood"), nullable=True
-        ),
-        sa.Column("created_at", sa.TIMESTAMP(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["care_receipient_id"],
-            ["care_receipient.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("mood")
-    op.drop_table("care_receipient")
-    op.drop_table("user")
-    # ### end Alembic commands ###
