+++ backend/alembic/versions/2026_01_04_1330-2bc13834fcef_init.py
@@ -1,0 +1,29 @@
+"""init
+
+Revision ID: 1f0eaf0d0ffe
+Revises:
+Create Date: 2026-01-04 13:30:01.225724
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "1f0eaf0d0ffe"
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    pass
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    pass
